+++ shadowsocksr-csharp/View/ConfigForm.cs
@@ -99,17 +99,17 @@ namespace ShadowsocksR.View
         {
             try
             {
-                return new Font("Consolas", 9F, FontStyle.Regular, GraphicsUnit.Point, ((byte)(0)));
+                return new Font("Consolas", 9F, FontStyle.Regular, GraphicsUnit.Point, 0);
             }
             catch
             {
                 try
                 {
-                    return new Font("微软雅黑", 9F, FontStyle.Regular, GraphicsUnit.Point, ((byte)(0)));
+                    return new Font("微软雅黑", 9F, FontStyle.Regular, GraphicsUnit.Point, 0);
                 }
                 catch
                 {
-                    return new Font(FontFamily.GenericMonospace, 9F, FontStyle.Regular, GraphicsUnit.Point, ((byte)(0)));
+                    return new Font(FontFamily.GenericMonospace, 9F, FontStyle.Regular, GraphicsUnit.Point, 0);
                 }
             }
         }
